pay-queue/requirements.txt
-e git+https://github.com/bcgov/sbc-common-components.git@5f99e135214ae949c9af951d4aa0b88b1067d853#egg=sbc_common_components&subdirectory=python
-e git+https://github.com/seeker25/sbc-pay.git@4846331b602fb56de171a32512b1578efc1489bb#egg=pay_api&subdirectory=pay-api
-e git+https://github.com/thorwolpert/flask-jwt-oidc.git@40cc811ccf70e838c5f7522fe8d83b7e58853539#egg=flask_jwt_oidc
CacheControl==0.14.0
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-OpenTracing==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Script==2.0.6
Flask==3.0.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
SQLAlchemy-Utils==0.41.1
SQLAlchemy==2.0.28
Werkzeug==3.0.1
alembic==1.13.1
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
attrs==23.2.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
croniter==2.0.2
cryptography==42.0.5
dpath==2.1.6
ecdsa==0.18.0
expiringdict==1.2.2
flask-marshmallow==1.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-pubsub==2.20.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio-status==1.62.1
grpcio==1.62.1
gunicorn==21.2.0
holidays==0.37
idna==3.6
itsdangerous==2.1.2
jaeger-client==4.8.0
jsonschema==4.17.3
launchdarkly-eventsource==1.1.1
launchdarkly-server-sdk==9.2.2
marshmallow-sqlalchemy==1.0.0
marshmallow==3.21.1
minio==7.2.5
msgpack==1.0.8
opentracing==2.4.0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyRFC3339==1.1
pyasn1-modules==0.3.0
pyasn1==0.5.1
pycountry==23.12.11
pycparser==2.21
pycryptodome==3.20.0
pyhumps==3.8.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
requests==2.31.0
rsa==4.9
semver==3.0.2
sentry-sdk==1.41.0
six==1.16.0
threadloop==1.0.2
thrift==0.16.0
tornado==6.4
typing_extensions==4.10.0
urllib3==2.2.1
